fx2lafw: Partial Revert of b99457f0
The original patch replaced a few cases where variables were
declared and initialised mid-function. Some of these were consts,
which were removed. Others were variable which could be made const
but have not been.

There have been some concerns raised about the readability of
variables/consts declared mid-way through functions. I believe
these are overblown. Certainly not of great enough significance
to warrent the wholesale removal all of these.

consts are an important protection offered by the C language, and
are a useful constraint for optimizers to exploit. They also
show new readers of the code that the variable is not reassigned
later, they also prevent later modifications to code from using
a variables in ways they were not intended to be used.

By definition a const can only be declared when it's value is
available - and this is often only possible mid-way though. This
is a legitimate reason to justify variables not always being
declared at the start of functions.

b99457f0 removed const declarations which this patch reintroduces.
It also removed variables decared mid-function - in each of these
latter cases these variables can be made const. A following patch
will make this change because any constraint that can be used
should be used, because stricter more tightly defined code is
always beneficial.